requirements.txt
numpy==2.2.6
pyarrow>=14.0.0
pandas>=2.2.0
tqdm
datasets>=2.20
evaluate>=0.4.2

torch==2.9.0
transformers==4.57.6
accelerate==1.12.0
trl==0.27.0
peft>=0.11.1
tokenizers==0.22.2
vllm==0.11.2

huggingface_hub
codebleu
orjson
matplotlib
bitsandbytes
python-dotenv
sentence-transformers
wandb>=0.17.0
